requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.65.0
pyyaml>=6.0
matplotlib>=3.7.0
plotly>=5.15.0

# NLP and clinical text processing
transformers>=4.30.0
datasets>=2.12.0
scispaCy>=0.5.3
medspacy>=1.0.0
sentencepiece>=0.1.99

# Evaluation and metrics
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.1

# Visualization and demo
streamlit>=1.25.0
gradio>=3.40.0

# Development and formatting
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
mypy>=1.4.0

# Optional: logging and tracking
wandb>=0.15.0
mlflow>=2.5.0

# Optional: serving
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
